basic impl
Implemented the following functions: 1. Player's movement and shooting, which adds the position of the gun. When the player switches direction, the gun will be aligned with the direction the player switches to facilitate the next step to achieve access to the handle. 2. added UI blood bar above the player, the player can pay attention to the battle situation at the same time, in time to notice their blood level. 3. optimized the code, made the encapsulation for the basic function, increase the code readability. And for the convenience of subsequent modifications.  4. Implemented the bullet shattering shatterable walls when dropping supplies. 5. Implemented two kinds of materials, rubies to increase blood, and sapphires to increase attack power.